Feature/324

See merge request s10-webmobile2-sub2/S10P12A408!97

+++ BE/src/main/java/com/ssafy/goumunity/domain/user/controller/request/UserModifyRequest.java
@@ -24,5 +24,5 @@ public class UserModifyRequest {
 
     private String nickname;
 
-    private Integer regionId;
+    private Long regionId;
 }
